// src/interfaces/Pokemon.tsx
export interface PokemonForm {
  name: string,
  id: number,
  image : string,
  type: string,
  abilities: string
}